# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/me": {
      "get": {
        "operationId": "UsersController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Return user profile.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get my profile",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/profile": {
      "put": {
        "operationId": "UsersController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User profile updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update my profile",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/public-profile/{id}": {
      "get": {
        "operationId": "UsersController_getPublicProfile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return public profile.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProfileDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        },
        "summary": "Get public profile by ID",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/projects/{projectId}/tasks": {
      "post": {
        "operationId": "TasksController_create",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new task under a project",
        "tags": [
          "Tasks"
        ]
      },
      "get": {
        "operationId": "TasksController_findAllByProject",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of tasks.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedTaskResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get all tasks by project ID",
        "tags": [
          "Tasks"
        ]
      }
    },
    "/tasks/public": {
      "get": {
        "operationId": "TasksController_getPublicTasks",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of public tasks.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedTaskResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get public tasks looking for assignees",
        "tags": [
          "Tasks"
        ]
      }
    },
    "/tasks/joined": {
      "get": {
        "operationId": "TasksController_getJoinedTasks",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of joined tasks.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedTaskResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get tasks assigned to me",
        "tags": [
          "Tasks"
        ]
      }
    },
    "/tasks/{id}": {
      "put": {
        "operationId": "TasksController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions to mark DONE"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a task",
        "tags": [
          "Tasks"
        ]
      }
    },
    "/tasks/{id}/comments": {
      "post": {
        "operationId": "TasksController_addComment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Comment added.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add a comment to a task",
        "tags": [
          "Tasks"
        ]
      },
      "get": {
        "operationId": "TasksController_getComments",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of comments.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommentResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all comments for a task",
        "tags": [
          "Tasks"
        ]
      }
    },
...
  },
  "info": {
    "title": "Task Bounty Platform API",
    "description": "The Nền tảng Quản lý công việc và Giao việc tự do API description.\n\n### Real-time Events (WebSocket)\nThis platform provides real-time updates via Socket.io.\n- **Endpoint**: `ws://<host>:<port>/notifications`\n- **Transports**: `['websocket', 'polling']`\n\n#### Events to Listen For\n1. **`risk-alert`**: Emitted when the AI detects a high risk score for a task.\n   - Payload: `{ pmId: string, message: string }`\n2. **`balance-warning`**: Emitted when the PM's escrow balance drops below a required threshold.\n   - Payload: `{ pmId: string, message: string }`\n",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UserProfileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-1234"
          },
          "email": {
            "type": "string",
            "example": "test@example.com"
          },
          "walletAddress": {
            "type": "string",
            "example": "0x123abc..."
          }
        },
        "required": [
          "id",
          "email",
          "walletAddress"
        ]
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "example": "password123"
          },
          "cvUrl": {
            "type": "string",
            "example": "https://example.com/cv.pdf"
          },
          "skills": {
            "example": [
              "React",
              "NestJS"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "bio": {
            "type": "string",
            "example": "Software Engineer"
          },
          "gender": {
            "type": "string",
            "example": "Male"
          },
          "birthYear": {
            "type": "number",
            "example": 1995
          },
          "contactInfo": {
            "type": "string",
            "example": "phone: 123456789"
          },
          "socialLinks": {
            "type": "object",
            "example": {
              "linkedin": "https://..."
            }
          }
        }
      },
      "PublicProfileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-1234"
          },
          "walletAddress": {
            "type": "string",
            "example": "0x123abc..."
          }
        },
        "required": [
          "id",
          "walletAddress"
        ]
      },
...
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "lastPage": {
            "type": "number"
          }
        },
        "required": [
          "total",
          "page",
          "limit",
          "lastPage"
        ]
      },
...
      "CreateTaskDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Implement login"
          },
          "description": {
            "type": "string",
            "example": "Implement the login flow using NestJS."
          },
          "budget": {
            "type": "number",
            "example": 500
          },
          "deadline": {
            "type": "string",
            "example": "2026-12-31T23:59:59Z"
          },
          "parentId": {
            "type": "string",
            "example": "parent-task-uuid"
          }
        },
        "required": [
          "title",
          "description"
        ]
      },
      "TaskResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-1234"
          },
          "projectId": {
            "type": "string",
            "example": "uuid-project-1234"
          },
          "title": {
            "type": "string",
            "example": "Implement login"
          },
          "description": {
            "type": "string",
            "example": "Implement the login flow using NestJS."
          },
          "status": {
            "type": "string",
            "example": "PENDING"
          },
          "budget": {
            "type": "number",
            "example": 500
          }
        },
        "required": [
          "id",
          "projectId",
          "title",
          "description",
          "status",
          "budget"
        ]
      },
      "PaginatedTaskResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskResponseDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "UpdateTaskDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "IN_PROGRESS",
            "enum": [
              "OPEN",
              "IN_PROGRESS",
              "REVIEW",
              "DONE"
            ]
          },
          "assigneeId": {
            "type": "string",
            "example": "uuid-of-assignee"
          },
          "description": {
            "type": "string",
            "example": "Updated description"
          }
        }
      },
      "CreateCommentDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "example": "Looking good!"
          }
        },
        "required": [
          "content"
        ]
      },
      "CommentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-1234"
          },
          "content": {
            "type": "string",
            "example": "Looking good!"
          },
          "userId": {
            "type": "string",
            "example": "uuid-user-1234"
          }
        },
        "required": [
          "id",
          "content",
          "userId"
        ]
      },
...
    }
  }
}